code_4_backend/requirements.txt
Django
numpy
pandas
matplotlib
scipy
djangorestframework
